updtaed code

+++ src/main/java/com/tejas/app/App.java
@@ -6,6 +6,6 @@ package com.tejas.app;
 public class App {
     public static void main(String[] args) {
         System.out.println("Hello World!");
-        int i = 20;
+        long i = 20;
     }
 }
